script auto commit

+++ ParsingScripts/MarksTableResult.php
@@ -126,12 +126,12 @@ class MarksTableTable {
         while ($i < $arrayLength)
         {
             $this->marksTable[$i] = array($this->weightings[$i], $this->denominators[$i], $this->emailNames[$i], $this->marks[$i]);
-            $cleanMark = (float)preg_replace("/[^0-9]+/", "", $this->marks[$i]);
-            $cleanDen = (float)preg_replace("/[^0-9]+/", "", $this->denominators[$i]);
+            $cleanMark = preg_replace("/[^0-9]+/", "", $this->marks[$i]);
+            $cleanDen = preg_replace("/[^0-9]+/", "", $this->denominators[$i]);
 
             // add percentage
             if ($cleanMark != "" && $cleanDen != "")
-                $this->marksTable[$i][4] = number_format((float)(($cleanMark / $cleanDen) * 100), 1, '.', '') . "%";
+                $this->marksTable[$i][4] = number_format((float)(((float)$cleanMark / (float)$cleanDen) * 100), 1, '.', '') . "%";
             else
                 $this->marksTable[$i][4] = "";
